+++ backend/app/configuration/environments/platform_service_configurations/schemas/postgresql_schema.py
@@ -1,0 +1,8 @@
+from .config_schema import BaseEnvironmentPlatformServiceConfigurationDetail
+
+
+class PostgreSQLConfig(BaseEnvironmentPlatformServiceConfigurationDetail):
+    host: str
+    port: str
+    admin_user: str
+    admin_pwd: str