+++ demo/basic/products/logistics_wms_shipments/models/shipments.sql
@@ -1,0 +1,7 @@
+SELECT
+    shipment_id,
+    order_ref,
+    shipped_date,
+    delivery_status
+FROM
+    {{ source('sources', 'wms_shipments') }}